restructure connections module directory for code clarity
