+++ omod/src/main/java/org/openmrs/module/patientportaltoolkit/fragment/controller/connections/AddRelationshipFragmentController.java
@@ -7,7 +7,7 @@
  * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
  * graphic logo is a trademark of OpenMRS Inc.
  */
-package org.openmrs.module.patientportaltoolkit.fragment.controller;
+package org.openmrs.module.patientportaltoolkit.fragment.controller.connections;
 
 import org.apache.commons.lang.StringUtils;
 import org.apache.commons.logging.Log;
